Refactor home page component into a few smaller components
